make the library more general by providing a content-type agnostic Tpl combinator and an HTML specific one. xss sanitize data that goes to HTML templates. the old Tpl is gone, since it was equivalent to Get '[HTML "tplfile"] Object
