+++ servant-ede.cabal
@@ -20,30 +20,31 @@ extra-source-files:  README.md
 cabal-version:       >=1.10
 
 library
-  exposed-modules:     Servant.EDE, Servant.EDE.HTML
+  exposed-modules:     Servant.EDE
   other-modules:
       Servant.EDE.Internal
     , Servant.EDE.Internal.Reify
-    , Servant.EDE.Internal.Templates
     , Servant.EDE.Internal.ToObject
     , Servant.EDE.Internal.Validate
 
   build-depends:
-      aeson >= 0.8
+      aeson >= 0.8 && <0.10
     , base >=4.7 && <5
-    , bytestring >= 0.10
-    , filepath >= 1.2
-    , ede >= 0.2
-    , either >= 4
-    , http-media >= 0.6
-    , http-types >= 0.8
-    , semigroups >= 0.16
-    , servant >= 0.4
-    , servant-server >= 0.4
-    , text >= 1.0
-    , transformers >= 0.4
-    , unordered-containers >= 0.2
-    , wai >= 3.0
+    , bytestring >= 0.10 && <0.11
+    , filepath >= 1.2 && <1.5
+    , ede >= 0.2 && <0.3
+    , either >= 4.0 && <4.5
+    , http-media >= 0.6 && <0.7
+    , http-types >= 0.8 && <0.9
+    , semigroups >= 0.16 && <0.17
+    , servant >= 0.4 && <0.5
+    , servant-server >= 0.4 && <0.5
+    , text >= 1.0 && <1.3
+    , transformers >= 0.4 && <0.5
+    , unordered-containers >= 0.2 && <0.3
+    , vector >= 0.9 && <0.11
+    , wai >= 3.0 && <3.1
+    , xss-sanitize >= 0.3 && <0.4
 
   ghc-options:         -O2 -Wall
   hs-source-dirs:      src
