+++ src/Servant/EDE/Internal.hs
@@ -1,0 +1,10 @@
+module Servant.EDE.Internal
+  ( module Servant.EDE.Internal.Reify
+  , module Servant.EDE.Internal.ToObject
+  , module Servant.EDE.Internal.Validate
+  ) where
+
+import Servant.EDE.Internal.Reify
+import Servant.EDE.Internal.ToObject
+import Servant.EDE.Internal.Validate
+
