--- src/Servant/HTML/EDE/Internal.hs
@@ -1,12 +1,0 @@
-module Servant.HTML.EDE.Internal
-  ( module Servant.HTML.EDE.Internal.Reify
-  , module Servant.HTML.EDE.Internal.Templates
-  , module Servant.HTML.EDE.Internal.ToObject
-  , module Servant.HTML.EDE.Internal.Validate
-  ) where
-
-import Servant.HTML.EDE.Internal.Reify
-import Servant.HTML.EDE.Internal.Templates
-import Servant.HTML.EDE.Internal.ToObject
-import Servant.HTML.EDE.Internal.Validate
-
